Added Module-Info, Javafx
Added module-info.java to the project in order to get JavaFX dependency
to run with Maven (javafx.graphics needs to be run in modulepath). Also
renamed the project to 'jmusicplayer' because the '-' was messing up the
module-info file. This version's App.java contains a GUI helloworld and
the SongHandler tests.