Audio, ListView, Pause/Resume
Audio Works with two underlying ListViews: One displays the Song._title,
one saves the Song object. Clicking the Song title in the ListView
selects the Song object and plays it that way.

Changed the Play and Pause buttons to ToggleButtons so when a song is
paused, hitting play resumes the song. Previously, hitting Play would
restart the song.